+++ ui/map/ImageOverlay.tsx
