+++ ui/map/Text.tsx
